doc:移动变量位置
